requirements2.txt
absl-py==1.4.0
addict==2.4.0
aliyun-python-sdk-core==2.16.0
aliyun-python-sdk-kms==2.16.5
appdirs==1.4.4
astunparse==1.6.3
beautifulsoup4==4.12.0
brotlipy==0.7.0
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
Connectivity==0.0.0
contourpy==1.0.7
crcmod==1.7
cryptography==38.0.4
cycler==0.11.0
docker-pycreds==0.4.0
filelock==3.14.0
flatbuffers==23.3.3
flit_core==3.6.0
fonttools==4.39.4
gast==0.4.0
gdown==4.7.1
gitdb==4.0.10
GitPython==3.1.30
google-auth==2.17.3
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.53.0
h5py==3.8.0
idna==3.4
imageio==2.27.0
importlib-metadata==6.4.1
importlib-resources==5.12.0
jax==0.4.8
jmespath==0.10.0
keras==2.11.0
kiwisolver==1.4.4
lazy_loader==0.2
libclang==16.0.0
Markdown==3.4.3
markdown-it-py==2.2.0
MarkupSafe==2.1.2
matplotlib==3.7.1
mdurl==0.1.2
mkl-fft==1.3.6
mkl-random==1.2.2
mkl-service==2.4.0
ml-dtypes==0.0.4
mmcv-full==1.7.1
mmdet==2.26.0
model-index==0.1.11
networkx==3.0
numpy==1.24.3
numpyencoder==0.3.0
oauthlib==3.2.2
opencv-python==4.7.0.72
opendatalab==0.0.10
openmim==0.3.7
openxlab==0.1.2
opt-einsum==3.3.0
ordered-set==4.1.0
oss2==2.17.0
packaging==23.1
pandas==2.0.1
pathtools==0.1.2
Pillow==9.4.0
pip==22.3.1
prettytable==3.6.0
protobuf==3.19.6
psutil==5.9.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0.6
pycparser==2.21
pycryptodome==3.21.0
Pygments==2.15.1
pyOpenSSL==22.0.0
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2023.3
PyWavelets==1.4.1
PyYAML==6.0
requests==2.28.2
requests-oauthlib==1.3.1
rich==13.3.5
rsa==4.9
scikit-image==0.20.0
scipy==1.9.1
sentry-sdk==1.14.0
setproctitle==1.3.2
setuptools==60.2.0
shapely==2.0.1
six==1.16.0
smmap==5.0.0
soupsieve==2.4
tabulate==0.9.0
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.11.0
tensorflow-addons==0.19.0
tensorflow-estimator==2.11.0
tensorflow-io-gcs-filesystem==0.32.0
termcolor==2.2.0
terminaltables==3.1.10
tifffile==2023.3.21
torch==1.13.1
torchaudio==0.13.1
torchvision==0.14.1
tqdm==4.65.0
typeguard==3.0.2
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.26.15
wandb==0.13.9
wcwidth==0.2.6
Werkzeug==2.2.3
wheel==0.37.1
wrapt==1.15.0
yapf==0.32.0
zipp==3.15.0
